Move `getrandom_syscall` from util_libc to `linux_android`.
