+++ src/linux_android_with_fallback.rs
@@ -1,10 +1,5 @@
 //! Implementation for Linux / Android with `/dev/urandom` fallback
-use crate::{
-    lazy::LazyBool,
-    linux_android, use_file,
-    util_libc::{getrandom_syscall, last_os_error},
-    Error,
-};
+use crate::{lazy::LazyBool, linux_android, use_file, util_libc::last_os_error, Error};
 use core::mem::MaybeUninit;
 
 pub fn getrandom_inner(dest: &mut [MaybeUninit<u8>]) -> Result<(), Error> {
@@ -18,7 +13,7 @@ pub fn getrandom_inner(dest: &mut [MaybeUninit<u8>]) -> Result<(), Error> {
 }
 
 fn is_getrandom_available() -> bool {
-    if getrandom_syscall(&mut []) < 0 {
+    if linux_android::getrandom_syscall(&mut []) < 0 {
         match last_os_error().raw_os_error() {
             Some(libc::ENOSYS) => false, // No kernel support
             // The fallback on EPERM is intentionally not done on Android since this workaround
